Fixed a bug in matrixio: The subroutine matio_spyMatrix did not work correctly for full matrices. This bug has been fixed.
Implemented a new feature in linearsystemscalar: lsyssc_MatMatMul.
This subroutine performs matrix-matrix multiplication for nearly all
combinations of matrix formats, e.g., full-diagonal, diagonal-full,
full-full, CRS-diagonal, diagonal-CRS, and CRS-CRS. The only
combination which is not supported is full-CRS or CRS-full since this
does not make too much sense. In general, the CRS format is used for
large matrices. If the multiplication with a full matrix is performed
a full (and very large) matrix will be generated.

In this revision, the algorithm for CRS-CRS is not yet finished. For
the time being, only the number of nonzero matrix entries resulting
from the multiplication C := A * B is computed. The symbolical and
numerical multiplication will follow ;-)