Included new feature for all(!!) supproted data types:
- The storage_getbase_XXX routines allow for the specification of
upper and lower bounds.

Example: CALL storage_getbase_int(h_Int,p_Int,80)

Here, the pointer p_Int is assigned the contents of h_Int but it is
restricted to the first 80 entries. Note, no checks are performed if
h_Int has enough memory. If this would be checked for each
getbase-call performance would be redruced. However, these low-level
routines should not be used by high-level application programmers. All
others just "know" what they are doing. 

For two-dimensional arrays only the lower/upper bounds of the second
dimension can be specified. This is consistent with the treatment of
reallocation, etc.

The main idea for these extended routines are as follows: In adaptive
procedures the required storage may vary from one step to the
other. However, reallocating the storage again and again is quite
costly. Hence, it makes sense to allocate memory which is "large
enough" throughout the simulation but only "use" some portion. This
is, what isprovided by the new getbase routines on the lowest level ;-)