Implemented resize mechanism for block vectors. Note that there is and will not be a resize mechanism for block matrices. This is due to the fact, that a block matrix is assembled from scalar matrices which can have arbitrary shape/format, etc. Hence, it does not make sence to reshape a complete block matrix.
Let me explain, how this resize mechanism for scalar/block vectors and
scalar matrices works and how it should be used. Normally, each of
these objects has a fixed shape which does not change. However, if you
think of adaptive grid refinement/coarsening the mesh changes
(possibly in each time step) and hence, all vectors/matrices need to
be adjusted. At first glance, each object could be released,
re-created in the correct size and filled with data. However, this is
prohibitively expensive if you think of time-dependent grid
adaptivity. Moreover, the complete reassembly of matrices is not
necessary but there exist techniques which require only local matrix
manipulations. As a starting point, vectors and matrices can be
"resized" on-the-fly. In short, if the available memory does not
suffice, than it needs to be reallocated physically which is quite
expensive. Suppose a vector has been allocated for NEQ entries but it
should be shrinked to NEQ-10. In this case the resize mechanism keeps
the reserved memory for NEQ entries and only adjusts the attribute NEQ
so that getbase returns the correct section of the memory. If in the
next step, the vector needs to be increased by 5, then no reallocates
has to be performed since NEQ+5 is still smaller than the initial size
of the vector.

Hopefully this explains which resize may be nice for your applications.