Tried to illustrate refinement/conversion routines by improving the ASCII-arts.
Added elements-meeting-at-vertex data structure to the adaptivity structure. This auxiliary structure is used in the recoarsening routine in order to update all neighbouring elements of a deleted/renumbered vertex. The elements-at-vertex structure is generate by the hadapt_genElementsAtVertex subroutine. During refinement/conversion it is updated continuously so as to reflect the current grid. The recoarsening algorithm is still to be implemented (I will do, soon!).