New features        : - Names of some structure variables shortened (RspatialDiscretisation -> RSpatialDiscr,...) Problems solved     : - MAJOR UPDATE. Discretisation structures invalid. Bugzilla-Bug-ID     : Additional comments :
The discretisation structures were not correct, the implementation of trial and test function association for bilinear forms in the structure was unusable for some PDE's! Now, the discretisation structure can accept only one element type, not multiple. To discretise a bilinear form, the createMatrixStructure allows to specify two different spatial discretisation structures for trial and test space. Both structures are now saved to the matrix to variables p_spatialDiscrTrial and p_spatialDiscrTest; they may of course coincide.