WARNING: This commit breaks the old build system using the hard-coded 'Makefile'.          I'm not sure about the Windows build system as I can not test it.
New features:

* parlst_getvalue_string_direct now expands all environment variables on-the-fly
  before returning.


Why does it break the old build system?
There is the Fortran 90/95 intrinsic GETENV which all Fortran 90/95 compilers I've seen
so far do provide. But GETENV does not provide whether or not an environment variable
is set. With GETENV it is impossible to distinguish the cases
* environment variable not set 
and
* environment variable set, but expands to an empty string
but refering to an unset environment variable in a dat file is supposed to cause a 
critical error while using an empty string in a dat file should clearly not.

Most compilers provide a more powerful version of GETENV: GET_ENVIRONMENT_VARIABLE.
Most, but not all. NEC compiler, PGI 6.1.x, PGI 6.2.x, 7.x don't.
By means of a preprecessor directive (#ifdef) the code is still portable.

But FEAT2's old build system does not support preprocessor directives.

Note: For the upcoming FEAT2 regression benchmark expansion of environment variables 
      in dat files is essential to keep the number of dat files at a reasonable and 
      managable size. And essential for it is the new build system, too.