FIXED: itc_initResidual did not set status to 'continue', thus iteration loops aborted before first iteration... ADDED: detection for NaNs in iteration control (if the compiler supports it)
Note:
The NaN detection requires the non-standard Fortran90 ISNAN() routine. The availability of this function is checked using an #ifdef HAS_INTRINSIC_ISNAN. The Visual Studio project files have been updates to predefine this macro, however, the un!x build system needs to define this macro when applicable. If the macro is undefined, NaN detection is disabled, but the code will compile than that. According to google, at least the GCC and IFC support the ISNAN routine.