ENH: added addition timings to solvers, especially multigrid
The t_linsolNode structure now contains two additional fields named 'dtimeInitStruct' and 'dtimeInitData', which measure the time for structure and data initialisation of each solver.
In addition to that, the 'Multigrid2' solver got several additional timing fields to measure the separate components used in multigrid, namely:
* Defect calculation time
* Pre-/Post-Smoothing times
* Coarse-Grid Solution time
* Prolongation/Restriction times
* Vector resorting time
* Defect filtering time
* Step Length Control time

The fields listed above are computed for the whole multigrid solver on one hand, as well as for each level separately. Currently there is no routine in the linearsolver module which summarises and prints out these timings (maybe added later), however, the timings can be read out by the application if desired.