FIX: Line integration routine used an improper cubature. ADD: Raytracing routine now returns the number of iterations.
WARNING: The line integration routine is not optimal. It does not go element by element but puts a line over all elements and applies cubature without respecting the element boundaries. Only for testing purposes. Should be rewritten later using a line search and computing the real number of elements on the way from the start to the end - or do a summed cubature element by element.