CHG: Code changed for new style of element distributions.
WARNING: This is a large update, but it does not really change underlying structures. Therefore, all applications should run without problems. The benchmark also works fine. However, if any error is encountered, please tell me immediately!!!
Michael


Explaination:
-------------
Element distributions do not exist in this form any more. Renamed to "element groups". An element distribution contained also cubature formulas etc. But as this did not make sense, cubature formulas in element distributions were made deprecated. But with this, the element distribution is nearly empty.

The element distribution is now replaced by "element groups". An element group is just a set of elements with the same element type. Kernel routines do not access the group structure anymore, all access is done via encapsuling subroutines (e.g. spdiscr_getElemGroupInfo).

In a second step, the element distributions will be removed completely. The new element groups will be realised by integer group sets which allow a faster access to the data.

In third step, a "trafo mesh" will be created which will be an extension to the standard triangulation, combined with a transformation formula. The trafo mesh will be realised by integer group sets as well, every group connected to a transformation formula. The element groups in the discretisation will be derived from the groups in the trafo mesh. Thus, each element group will be connected to a transformation formula, an the trafo mesh contains the data how to apply the transformation.

The trafo mesh will then be designed in such a way that it supports isoparametric mappings.

Finally, the block assembly routines will be extended to support the extended transformation. That way it should be possible to include isoparametric elements.