updated VS thirdparty build-system to SuiteSparse
The third-party download and build scripts for Visual Studio have been adjusted to the new download locations of SuiteSparse, as the old links to AMD, UMFPACK and SuiteSparse_config are orphaned by now.

Note #1:
This commit only affects the VS+IFC build system for Windows; the linux/un!x build system has been patched already in rev 12162.

Note #2:
Users which already have a configured windows checkout do not necessarily need to re-download and re-compile the third-party libraries, as this commit is backwards compatible.